+++ memory-serve-core/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "memory-serve-core"
+description = "Shared code for memory-serve and memory-serve-macros"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+publish.workspace = true
+
+[dependencies]
+sha256 = "1.4"
+brotli = "7.0"
+mime_guess = "2.0"
+walkdir = "2"
