+++ memory-serve/src/lib.rs
@@ -9,9 +9,13 @@ mod asset;
 mod cache_control;
 mod util;
 
+#[allow(unused)]
+use crate as memory_serve;
+
 use crate::util::{compress_gzip, decompress_brotli};
 
 pub use crate::{asset::Asset, cache_control::CacheControl};
+pub use memory_serve_macros::load_assets;
 
 #[derive(Debug, Clone, Copy)]
 struct ServeOptions {
@@ -52,10 +56,17 @@ pub struct MemoryServe {
 }
 
 impl MemoryServe {
+    pub fn new(assets: &'static [Asset]) -> Self {
+        Self {
+            assets,
+            ..Default::default()
+        }
+    }
+
     /// Initiate a `MemoryServe` instance, takes the contents of `memory_serve_assets.bin`
     /// created at build time.
     /// Specify which asset directory to include using the environment variable `ASSET_DIR`.
-    pub fn new() -> Self {
+    pub fn from_env() -> Self {
         let assets: &[(&str, &[Asset])] =
             include!(concat!(env!("OUT_DIR"), "/memory_serve_assets.rs"));
 
@@ -63,10 +74,7 @@ impl MemoryServe {
             panic!("No assets found, did you forget to set the ASSET_DIR environment variable?");
         }
 
-        Self {
-            assets: assets[0].1,
-            ..Default::default()
-        }
+        Self::new(assets[0].1)
     }
 
     /// Include a directory using a named environment variable, prefixed by ASSRT_DIR_.
@@ -74,7 +82,7 @@ impl MemoryServe {
     /// The name should be in uppercase.
     /// For example to include assets from the public directory using the name PUBLIC, set the enirobment variable
     /// `ASSET_DIR_PUBLIC=./public` and call `MemoryServe::from_name("PUBLIC")`.
-    pub fn from_name(name: &str) -> Self {
+    pub fn from_env_name(name: &str) -> Self {
         let assets: &[(&str, &[Asset])] =
             include!(concat!(env!("OUT_DIR"), "/memory_serve_assets.rs"));
 
@@ -90,10 +98,7 @@ impl MemoryServe {
             );
         }
 
-        Self {
-            assets,
-            ..Default::default()
-        }
+        Self::new(assets)
     }
 
     /// Which static file to serve on the route "/" (the index)
@@ -288,7 +293,7 @@ impl MemoryServe {
 
 #[cfg(test)]
 mod tests {
-    use crate::{Asset, CacheControl, MemoryServe};
+    use crate::{self as memory_serve, Asset, CacheControl, MemoryServe};
     use axum::{
         body::Body,
         http::{
@@ -298,6 +303,7 @@ mod tests {
         },
         Router,
     };
+    use memory_serve_macros::load_assets;
     use tower::ServiceExt;
 
     async fn get(
@@ -325,11 +331,9 @@ mod tests {
         headers.get(name).unwrap().to_str().unwrap()
     }
 
-    #[test]
-    fn test_load_assets() {
-        let assets_list: &[(&str, &[Asset])] =
-            include!(concat!(env!("OUT_DIR"), "/memory_serve_assets.rs"));
-        let assets = assets_list[0].1;
+    #[tokio::test]
+    async fn test_load_assets() {
+        let assets: &[Asset] = load_assets!("../static");
         let routes: Vec<&str> = assets.iter().map(|a| a.route).collect();
         let content_types: Vec<&str> = assets.iter().map(|a| a.content_type).collect();
         let etags: Vec<&str> = assets.iter().map(|a| a.etag).collect();
@@ -378,7 +382,7 @@ mod tests {
 
     #[tokio::test]
     async fn if_none_match_handling() {
-        let memory_router = MemoryServe::new().into_router();
+        let memory_router = MemoryServe::new(load_assets!("../static")).into_router();
         let (code, headers) =
             get(memory_router.clone(), "/index.html", "accept", "text/html").await;
         let etag: &str = headers.get(header::ETAG).unwrap().to_str().unwrap();
@@ -398,7 +402,7 @@ mod tests {
 
     #[tokio::test]
     async fn brotli_compression() {
-        let memory_router = MemoryServe::new().enable_brotli(true).into_router();
+        let memory_router = MemoryServe::new(load_assets!("../static")).enable_brotli(true).into_router();
         let (code, headers) = get(
             memory_router.clone(),
             "/index.html",
@@ -414,7 +418,7 @@ mod tests {
         assert_eq!(length.parse::<i32>().unwrap(), 178);
 
         // check disable compression
-        let memory_router = MemoryServe::new().enable_brotli(false).into_router();
+        let memory_router = MemoryServe::new(load_assets!("../static")).enable_brotli(false).into_router();
         let (code, headers) = get(
             memory_router.clone(),
             "/index.html",
@@ -430,7 +434,7 @@ mod tests {
 
     #[tokio::test]
     async fn gzip_compression() {
-        let memory_router = MemoryServe::new().enable_gzip(true).into_router();
+        let memory_router = MemoryServe::new(load_assets!("../static")).enable_gzip(true).into_router();
         let (code, headers) = get(
             memory_router.clone(),
             "/index.html",
@@ -447,7 +451,7 @@ mod tests {
         assert_eq!(length.parse::<i32>().unwrap(), 274);
 
         // check disable compression
-        let memory_router = MemoryServe::new().enable_gzip(false).into_router();
+        let memory_router = MemoryServe::new(load_assets!("../static")).enable_gzip(false).into_router();
         let (code, headers) = get(
             memory_router.clone(),
             "/index.html",
@@ -463,12 +467,12 @@ mod tests {
 
     #[tokio::test]
     async fn index_file() {
-        let memory_router = MemoryServe::new().index_file(None).into_router();
+        let memory_router = MemoryServe::new(load_assets!("../static")).index_file(None).into_router();
 
         let (code, _) = get(memory_router.clone(), "/", "accept", "*").await;
         assert_eq!(code, 404);
 
-        let memory_router = MemoryServe::new()
+        let memory_router = MemoryServe::new(load_assets!("../static"))
             .index_file(Some("/index.html"))
             .into_router();
 
@@ -478,7 +482,7 @@ mod tests {
 
     #[tokio::test]
     async fn index_file_on_subdirs() {
-        let memory_router = MemoryServe::new()
+        let memory_router = MemoryServe::new(load_assets!("../static"))
             .index_file(Some("/index.html"))
             .index_on_subdirectories(false)
             .into_router();
@@ -486,7 +490,7 @@ mod tests {
         let (code, _) = get(memory_router.clone(), "/blog", "accept", "*").await;
         assert_eq!(code, 404);
 
-        let memory_router = MemoryServe::new()
+        let memory_router = MemoryServe::new(load_assets!("../static"))
             .index_file(Some("/index.html"))
             .index_on_subdirectories(true)
             .into_router();
@@ -497,7 +501,7 @@ mod tests {
 
     #[tokio::test]
     async fn clean_url() {
-        let memory_router = MemoryServe::new().enable_clean_url(true).into_router();
+        let memory_router = MemoryServe::new(load_assets!("../static")).enable_clean_url(true).into_router();
 
         let (code, _) = get(memory_router.clone(), "/about.html", "accept", "*").await;
         assert_eq!(code, 404);
@@ -508,11 +512,11 @@ mod tests {
 
     #[tokio::test]
     async fn fallback() {
-        let memory_router = MemoryServe::new().into_router();
+        let memory_router = MemoryServe::new(load_assets!("../static")).into_router();
         let (code, _) = get(memory_router.clone(), "/foobar", "accept", "*").await;
         assert_eq!(code, 404);
 
-        let memory_router = MemoryServe::new()
+        let memory_router = MemoryServe::new(load_assets!("../static"))
             .fallback(Some("/index.html"))
             .into_router();
         let (code, headers) = get(memory_router.clone(), "/foobar", "accept", "*").await;
@@ -520,7 +524,7 @@ mod tests {
         assert_eq!(code, 404);
         assert_eq!(length.parse::<i32>().unwrap(), 437);
 
-        let memory_router = MemoryServe::new()
+        let memory_router = MemoryServe::new(load_assets!("../static"))
             .fallback(Some("/index.html"))
             .fallback_status(StatusCode::OK)
             .into_router();
@@ -533,7 +537,7 @@ mod tests {
     #[tokio::test]
     async fn cache_control() {
         async fn check_cache_control(cache_control: CacheControl, expected: &str) {
-            let memory_router = MemoryServe::new()
+            let memory_router = MemoryServe::new(load_assets!("../static"))
                 .cache_control(cache_control)
                 .into_router();
 
@@ -567,7 +571,7 @@ mod tests {
         .await;
 
         async fn check_html_cache_control(cache_control: CacheControl, expected: &str) {
-            let memory_router = MemoryServe::new()
+            let memory_router = MemoryServe::new(load_assets!("../static"))
                 .html_cache_control(cache_control)
                 .into_router();
 
@@ -601,7 +605,7 @@ mod tests {
 
     #[tokio::test]
     async fn aliases() {
-        let memory_router = MemoryServe::new()
+        let memory_router = MemoryServe::new(load_assets!("../static"))
             .add_alias("/foobar", "/index.html")
             .add_alias("/baz", "/index.html")
             .into_router();
