+++ memory-serve/src/util.rs
